mortgage_broker_api/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.0
anyio==3.6.2
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
build==0.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.2
django-allauth==0.54.0
django-environ==0.10.0
django-oauth-toolkit==2.3.0
django-oauth2-provider==0.2.6.1
django-rest-authtoken==2.1.4
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docker-pycreds==0.4.0
drf-social-oauth2==2.1.3
duckdb==0.8.0
ecdsa==0.18.0
faiss-cpu==1.7.4
fastapi==0.95.2
ffmpy==0.3.0
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.31
google-search-results==2.4.2
gpt-index==0.4.24
gradio==3.32.0
gradio_client==0.2.5
greenlet==2.0.2
h11==0.14.0
hiredis==2.2.3
hnswlib==0.7.0
httpcore==0.17.1
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jwcrypto==1.5.0
kiwisolver==1.4.4
langchain==0.0.154
linkify-it-py==2.0.2
llama-index==0.6.9
loguru==0.7.0
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.15.0
openai==0.27.6
openapi-schema-pydantic==1.2.4
orjson==3.8.12
overrides==7.3.1
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.5.0
pinecone-client==2.2.2
pip-tools==6.13.0
posthog==3.0.1
protobuf==4.23.1
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.6
pulsar-client==3.2.0
py==1.11.0
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==3.5.3
redisearch==2.1.1
regex==2023.5.5
rejson==0.5.6
requests==2.29.0
requests-oauthlib==1.3.1
retry==0.9.2
rsa==4.9
semantic-version==2.10.0
sentry-sdk==1.23.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SQLAlchemy==1.4.48
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.15
uvicorn==0.22.0
wandb==0.15.3
watchfiles==0.19.0
websockets==11.0.3
win32-setctime==1.1.0
wrapt==1.15.0
yarl==1.9.2
zstandard==0.21.0
